requirements.txt
Django==3.1.5
djangorestframework==3.12.2
django-filter==2.4.0
Markdown==3.3.4
#psycopg2==2.8.6
